Merge mandatory/mydata, lots of other cleanup.
I'm working on bringing this into the modern era. Much of this work
is just simple housecleaning and removing the "old way" The plan
is to simplify the networking and just use wireguard, instead of the earlier
version which used hamachi and some tunneling programs for backup.

Another change is considering syncthing for backing up data, though
this may prove difficult.

All the old wrappers are gone, and now we have a python script that
runs arecord for pull data from the audio inputs.

Another script is in charge of deleting old data, ensuring that it only
takes up a specific amount of space.